Move ThunderBaseTest to BTB
