+++ src/ThunderTestTrait.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Drupal\thunder;
+
+use Drupal\Component\Utility\Crypt;
+use Drupal\Core\Database\Database;
+use Symfony\Component\HttpFoundation\Request;
+
+/**
+ * Use this trait to reuse an existing database.
+ */
+trait ThunderTestTrait {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function installDrupal() {
+    $this->initUserSession();
+    $this->prepareSettings();
+    $this->doInstall();
+    $this->initSettings();
+    $request = Request::createFromGlobals();
+    $container = $this->initKernel($request);
+    $this->initConfig($container);
+    $this->installModulesFromClassProperty($container);
+    $this->rebuildAll();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function prepareSettings() {
+    parent::prepareSettings();
+
+    // Remember the profile which was used.
+    $settings['settings']['install_profile'] = (object) [
+      'value' => $this->profile,
+      'required' => TRUE,
+    ];
+    // Generate a hash salt.
+    $settings['settings']['hash_salt'] = (object) [
+      'value'    => Crypt::randomBytesBase64(55),
+      'required' => TRUE,
+    ];
+
+    // Since the installer isn't run, add the database settings here too.
+    $settings['databases']['default'] = (object) [
+      'value' => Database::getConnectionInfo(),
+      'required' => TRUE,
+    ];
+
+    $this->writeSettings($settings);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function doInstall() {
+
+    if (!empty($_SERVER['thunderDumpFile'])) {
+      $file = $_SERVER['thunderDumpFile'];
+      // Load the database.
+      if (substr($file, -3) == '.gz') {
+        $file = "compress.zlib://$file";
+      }
+      require $file;
+    }
+    else {
+      parent::doInstall();
+    }
+  }
+
+}
